+++ src/jigsaw/pc-components/box/box.scss
@@ -6,11 +6,12 @@ $flex-prefix-cls: #{$jigsaw-prefix}-flex;
     display: block;
     flex: 1 0 0;
     overflow: auto;
-    &-relative {
+    &-resize-line-wrapper {
         position: relative;
         display: block;
         width: 0;
         height: 0;
+        pointer-events: auto;
     }
     &-resize-line {
         position: absolute;
